@@ -3,14 +3,14 @@ version: 0.5.2
 description: Android utilities
 publish_to: none
 environment:
-  sdk: '>=2.10.0-0 <3.0.0'
+  sdk: '>=2.12.0 <3.0.0'
 dependencies:
   args: any
   fs_shim: '>=0.7.1'
   tekartik_io_auth_utils:
     git:
       url: git://github.com/tekartik/io_auth_utils.dart
-      ref: dart2
+      ref: null_safety
   #tekartik_build_utils:
   #  git:
   #    url: git://github.com/tekartik/build_utils.dart
@@ -19,7 +19,7 @@ dependencies:
   tekartik_cmd_utils:
     git:
       url: git://github.com/tekartik/cmd_utils.dart
-      ref: dart2
+      ref: null_safety
     version: '>=0.2.2'
   googleapis_auth: any
   googleapis: any
