@@ -1,7 +1,7 @@
 class ApkInfo {
-  String name;
-  String versionName;
-  String versionCode;
+  String? name;
+  String? versionName;
+  String? versionCode;
 
   ApkInfo({this.name, this.versionName, this.versionCode});
 
