@@ -1,5 +1,5 @@
 import 'package:tekartik_android_utils/src/shell/shell.dart';
 
-String _gradleExecutableFilename;
+String? _gradleExecutableFilename;
 String get gradleShellExecutableFilename =>
     _gradleExecutableFilename ??= getBashOrBatExecutableFilename('gradlew');
